requirements.txt
pandas>=2.1.0
openpyxl>=3.1.2
pydantic>=2.5.0
langgraph>=0.2.0
openai>=1.55.0
python-dotenv>=1.0.0
streamlit>=1.30.0
